+++ Nzy3d/Colors/IColorMappable.cs
@@ -5,18 +5,11 @@ namespace Nzy3d.Colors
 		/// <summary>
 		/// Get/Set the lower value boundary for a <see cref="ColorMaps.IColorMap"/>.
 		/// </summary>
-		double ZMin { get; set; }
+		float ZMin { get; set; }
 
-        /// <summary>
-        /// Get/Set the upper value boundary for a <see cref="ColorMaps.IColorMap"/>.
-        /// </summary>
-        double ZMax { get; set; }
+		/// <summary>
+		/// Get/Set the upper value boundary for a <see cref="ColorMaps.IColorMap"/>.
+		/// </summary>
+		float ZMax { get; set; }
 	}
 }
-
-//=======================================================
-//Service provided by Telerik (www.telerik.com)
-//Conversion powered by NRefactory.
-//Twitter: @telerik
-//Facebook: facebook.com/telerik
-//=======================================================
