+++ Nzy3d/Maths/Normal.cs
@@ -7,7 +7,7 @@ namespace Nzy3d.Maths
 			var v1 = new Vector3d(p0, p1);
 			var v2 = new Vector3d(p1, p2);
 			var norm = v1.Cross(v2);
-			double d = norm.Distance(Coord3d.ORIGIN);
+			float d = norm.Distance(Coord3d.ORIGIN);
 			return norm.Divide(d);
 		}
 	}
