+++ Nzy3d/Plot3D/Builder/Concrete/RingGrid.cs
@@ -6,27 +6,27 @@ namespace Nzy3d.Plot3D.Builder.Concrete
     public class RingGrid : Grid
 	{
 		internal double sqradius;
-		public RingGrid(double radius, int xysteps, int enlargeSteps) : base(new Range(-radius - (enlargeSteps * radius / xysteps), radius + (enlargeSteps * radius / xysteps)), xysteps)
+		public RingGrid(float radius, int xysteps, int enlargeSteps) : base(new Range(-radius - (enlargeSteps * radius / xysteps), radius + (enlargeSteps * radius / xysteps)), xysteps)
 		{
 			sqradius = (radius + (enlargeSteps * radius / xysteps)) * (radius + (enlargeSteps * radius / xysteps));
 		}
 
-		public RingGrid(double radius, int xysteps) : this(radius, xysteps, 0)
+		public RingGrid(float radius, int xysteps) : this(radius, xysteps, 0)
 		{
 		}
 
 		public override List<Coord3d> Apply(Mapper mapper)
 		{
-			double xstep = xrange.Range / xsteps;
-			double ystep = yrange.Range / ysteps;
+			float xstep = xrange.Range / xsteps;
+			float ystep = yrange.Range / ysteps;
 			var output = new List<Coord3d>();
 
 			for (int xi = -(xsteps - 1) / 2; xi <= (xsteps - 1) / 2; xi++)
 			{
 				for (int yi = -(ysteps - 1) / 2; yi <= (ysteps - 1) / 2; yi++)
 				{
-					double x = xi * xstep;
-					double y = yi * ystep;
+					float x = xi * xstep;
+					float y = yi * ystep;
 					if (sqradius > x * x + y * y)
 					{
 						output.Add(new Coord3d(x, y, mapper.f(x, y)));
